fehler behoben

+++ src/de/tuberlin/cit/vs/applications/BillingSystem.java
@@ -8,6 +8,8 @@ import org.apache.camel.Exchange;
 import org.apache.camel.Processor;
 import org.apache.camel.builder.RouteBuilder;
 import org.apache.camel.impl.DefaultCamelContext;
+import org.apache.camel.util.FileUtil;
+
 
 import javax.jms.*;
 import java.io.BufferedWriter;
@@ -20,17 +22,23 @@ public class BillingSystem {
 
 
     public static final String billingPath = "data/billing_validated_orders/billing_validation.txt";
-    public static final String validatedPath = "data/validated_orders/validated_orders.txt";
+    public static final String validatedPath = "data/validated_orders";
     public static final String ackPath = "data/ACK.txt";
     public static int count = 0;
 
 
     static {
-        File file = new File(billingPath);
-        file.delete();
+        //delete billingOrders before start
+        File file1 = new File(billingPath);
+        file1.delete();
 
+        //delete every validated order before start
         File file2 = new File(validatedPath);
-        file2.delete();
+        for(File file: file2.listFiles())
+            if (!file.isDirectory())
+                file.delete();
+
+
     }
 
     public static void writeToFile(String path, String content, Boolean append) throws IOException {
@@ -42,7 +50,9 @@ public class BillingSystem {
 
     private static void updateFile(int count, String validatedOrder) throws IOException {
         //wait till billingOrder has finished
+        //improvised synchronisation between inventorySystem and billingSystem
         while (true) {
+            //read ACK.txt file until InventorySystem sends an ACK: OKFromInventory
             int condition = 0;
             File file = new File(ackPath);
             Scanner sc = new Scanner(file);
@@ -57,6 +67,7 @@ public class BillingSystem {
         }
 
         writeToFile("data/validated_orders/validated_orders_" + count + ".txt", validatedOrder, true);
+        //sending own ACK: OKFromBilling
         writeToFile(ackPath, "OKFromBilling", false);
 
 
@@ -69,10 +80,9 @@ public class BillingSystem {
             conFactory.setTrustAllPackages(true);
             Connection con = conFactory.createConnection();
 
-            // created two queues for getting and sending messages back
+            // listen to the topic: billingIn
             final Session session = con.createSession(false, Session.AUTO_ACKNOWLEDGE);
             Topic billingIn = session.createTopic("billingIn");
-            Queue billingOut = session.createQueue("billingOut");
             MessageConsumer consumer = session.createConsumer(billingIn);
             consumer.setMessageListener(new MessageListener() {
 
@@ -85,22 +95,19 @@ public class BillingSystem {
                         order.validate();
 
                         // randomly decide whether customer has enough money
-                        ObjectMessage answer = session.createObjectMessage(order);
                         boolean payed = Math.random() > 0.5;
-                        answer.setBooleanProperty("payed", payed);
                         if(payed) order.setValid("1");
-                        answer.setBooleanProperty("available", false);
 
-
-
+                        //writes the order to billing_validation.txt
                         String validatedOrder = order.getOrderForCall() + "," + order.getValid() + ",0\n";
 
                         writeToFile(billingPath, validatedOrder, true);
+                        //synchronizes with inventorySystem
                         updateFile(BillingSystem.count, validatedOrder);
                         BillingSystem.count++;
 
-                        //MessageProducer producer = session.createProducer(billingOut);
-                        //producer.send(answer);
+
+
                     } catch (JMSException | IOException e) {
                         e.printStackTrace();
                     }
